+++ logic_if_else.go
@@ -49,8 +49,11 @@ import "io"
 func IfElse(expr, body, elseBlock *Sequence) Command {
 	// build our Scriptish Command
 	return func(p *Pipe) (int, error) {
+		// get our parameters
+		params := getParamsFromEnv(p.Env)
+
 		// run the test expression first
-		expr.Exec()
+		expr.Exec(params...)
 
 		// copy the output over to our pipe
 		io.Copy(p.Stdout, expr.Pipe.Stdout.NewReader())
@@ -60,7 +63,7 @@ func IfElse(expr, body, elseBlock *Sequence) Command {
 		err := expr.Error()
 		if err == nil {
 			// yes we can!
-			body.Exec()
+			body.Exec(params...)
 
 			// copy the output over to our pipe
 			io.Copy(p.Stdout, body.Pipe.Stdout.NewReader())